requirements.txt
awsebcli==3.14.6
blessed==1.15.0
botocore==1.12.23
cached-property==1.5.1
cement==2.8.2
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
oauthlib==2.1.0
pathspec==0.5.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.0.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.53.0
Pillow==5.3.0
